+++ contracts/container/contract.go
@@ -106,8 +106,11 @@ const (
 	nnsRootKey         = "nnsRoot"
 	nnsHasAliasKey     = "nnsHasAlias"
 
-	corsAttributeName = "CORS"
-	lockAttributeName = "__NEOFS__LOCK_UNTIL"
+	corsAttributeName            = "CORS"
+	lockAttributeName            = "__NEOFS__LOCK_UNTIL"
+	s3TagsAttributeName          = "S3_TAGS"
+	s3SettingsAttributeName      = "S3_SETTINGS"
+	s3NotificationsAttributeName = "S3_NOTIFICATIONS"
 
 	// nolint:unused
 	nnsDefaultTLD = "container"
@@ -2364,12 +2367,25 @@ func attributeFromBytes(b []byte) (Attribute, string) {
 // with SetAttribute. The supported list of attributes:
 //   - CORS
 //   - NEOFS__LOCK_UNTIL
+//   - S3_TAGS
+//   - S3_SETTINGS
+//   - S3_NOTIFICATIONS
 //
 // CORS attribute gets JSON encoded `[]CORSRule` as value.
 //
 // If name is '__NEOFS__LOCK_UNTIL', value must a valid Unix Timestamp later the
 // current and already set (if any) ones. On success, referenced container
 // becomes locked for removal until specified time.
+//
+// If the name is 'S3_TAGS', the value must be a valid JSON map, where the key is the tag name and
+// the value is the tag value. It is an S3 gate-specific attribute.
+// For instance: {"my-tag":"my-value"}.
+//
+// If the name is 'S3_SETTINGS', the value is not validated by the contract, but must be valid JSON.
+// It is an S3 gate-specific attribute. The structure of it is controlled by the gate itself.
+//
+// If the name is 'S3_NOTIFICATIONS', the value is not validated by the contract, but must be valid JSON.
+// It is an S3 gate-specific attribute. The structure of it is controlled by the gate itself.
 //
 // The validUntil must be Unix Timestamp which has not yet pass.
 //
@@ -2421,6 +2437,20 @@ func SetAttribute(cID interop.Hash256, name, value string, validUntil int, invoc
 		if idx >= 0 && until <= std.Atoi10(info.Attributes[idx].Value) {
 			panic("lock expiration time " + value + " is not later than already set " + info.Attributes[idx].Value)
 		}
+	case s3TagsAttributeName:
+		tags := std.JSONDeserialize([]byte(value)).(map[string]any)
+		for k, v := range tags {
+			if k == "" {
+				panic("tag key is empty")
+			}
+			if v.(string) == "" {
+				panic("tag " + k + " value is empty")
+			}
+		}
+	case s3SettingsAttributeName:
+		_ = std.JSONDeserialize([]byte(value)).(map[string]any)
+	case s3NotificationsAttributeName:
+		_ = std.JSONDeserialize([]byte(value)).(map[string]any)
 	default:
 		panic("attribute is immutable")
 	}
@@ -2489,7 +2519,7 @@ func validateCORSRule(rule map[string]any) string {
 		return err
 	}
 
-	if maxAgeSeconds := rule["MaxAgeSeconds"].(int64); maxAgeSeconds < 0 {
+	if maxAgeSeconds, ok := rule["MaxAgeSeconds"]; ok && maxAgeSeconds.(int64) < 0 {
 		return "MaxAgeSeconds must be >= 0"
 	}
 
@@ -2536,10 +2566,6 @@ func validateCORSAllowedOrigins(items []any) string {
 }
 
 func validateCORSAllowedHeaders(items []any) string {
-	if len(items) == 0 {
-		return "AllowedHeaders is empty"
-	}
-
 	for i, header := range items {
 		h := header.(string)
 		if h == "" {
@@ -2570,6 +2596,9 @@ func validateCORSExposeHeaders(items []any) string {
 // with RemoveAttribute. The supported list of attributes:
 //   - CORS
 //   - __NEOFS__LOCK_UNTIL
+//   - S3_TAGS
+//   - S3_SETTINGS
+//   - S3_NOTIFICATIONS
 //
 // If name is '__NEOFS__LOCK_UNTIL', current time must be later than the
 // currently set one if any.
@@ -2610,6 +2639,9 @@ func RemoveAttribute(cID interop.Hash256, name string, validUntil int, invocScri
 
 	switch name {
 	case corsAttributeName:
+	case s3TagsAttributeName:
+	case s3SettingsAttributeName:
+	case s3NotificationsAttributeName:
 	case lockAttributeName:
 		if index >= 0 {
 			now := runtime.GetTime()