+++ tests/container_test.go
@@ -3447,19 +3447,18 @@ func TestSetCORSAttribute(t *testing.T) {
 	})
 
 	t.Run("allowedHeaders", func(t *testing.T) {
-		t.Run("empty headers list", func(t *testing.T) {
+		t.Run("empty headers list ok", func(t *testing.T) {
 			var rules = []CORSRule{
 				{
 					AllowedMethods: []string{"GET"},
 					AllowedOrigins: []string{"*"},
-					AllowedHeaders: []string{},
 				},
 			}
 
 			pl, err := json.Marshal(rules)
 			require.NoError(t, err)
 
-			inv.InvokeFail(t, "AllowedHeaders is empty", "setAttribute", cID[:], "CORS", pl,
+			inv.Invoke(t, nil, "setAttribute", cID[:], "CORS", pl,
 				anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
 		})
 
@@ -3590,5 +3589,168 @@ func TestSetCORSAttribute(t *testing.T) {
 
 			assertGetInfo(t, inv, cID, cnr)
 		})
+	})
+}
+
+func TestSetS3TAGSAttribute(t *testing.T) {
+	const anyValidUntil = math.MaxUint32
+	anyValidInvocScript := randomBytes(10)
+	anyValidVerifScript := randomBytes(10)
+	anyValidSessionToken := randomBytes(10)
+
+	blockChain, committee := chain.NewSingle(t)
+	require.Implements(t, (*neotest.MultiSigner)(nil), committee)
+	exec := neotest.NewExecutor(t, blockChain, committee, committee)
+
+	deployDefaultNNS(t, exec)
+	netmapContract := deployNetmapContract(t, exec, "ContainerFee", 0)
+	containerContract := neotest.CompileFile(t, exec.CommitteeHash, containerPath, path.Join(containerPath, "config.yml"))
+	deployBalanceContract(t, exec, netmapContract, containerContract.Hash)
+	deployProxyContract(t, exec)
+
+	exec.DeployContract(t, containerContract, nil)
+
+	owner := exec.NewAccount(t, 200_0000_0000)
+	ownerAcc := owner.ScriptHash()
+	ownerAddr := user.NewFromScriptHash(ownerAcc)
+	cnr := containertest.Container()
+	cnr.SetOwner(ownerAddr)
+	cnrBytes := cnr.Marshal()
+	cID := cid.NewFromMarshalledContainer(cnrBytes)
+	inv := exec.CommitteeInvoker(containerContract.Hash)
+
+	t.Run("create container", func(t *testing.T) {
+		const anyValidDomainName = ""
+		const anyValidDomainZone = ""
+
+		committeeInvoker := exec.CommitteeInvoker(containerContract.Hash)
+
+		_ = committeeInvoker.Invoke(t, stackitem.Null{}, "create",
+			cnrBytes, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken, anyValidDomainName, anyValidDomainZone, false)
+
+		assertGetInfo(t, inv, cID, cnr)
+	})
+
+	t.Run("store and remove S3_TAGS", func(t *testing.T) {
+		var tags = map[string]string{
+			"key": "value",
+		}
+
+		pl, err := json.Marshal(tags)
+		require.NoError(t, err)
+
+		var cnr2 container.Container
+		cnr.CopyTo(&cnr2)
+
+		inv.Invoke(t, nil, "setAttribute", cID[:], "S3_TAGS", pl, anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+		cnr2.SetAttribute("S3_TAGS", string(pl))
+		assertGetInfo(t, inv, cID, cnr2)
+
+		inv.Invoke(t, nil, "removeAttribute", cID[:], "S3_TAGS", anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+		assertGetInfo(t, inv, cID, cnr)
+	})
+
+	t.Run("key is empty", func(t *testing.T) {
+		var tags = map[string]string{
+			"k1": "value1",
+			"":   "value2",
+			"k3": "value3",
+		}
+
+		pl, err := json.Marshal(tags)
+		require.NoError(t, err)
+
+		inv.InvokeFail(t, "tag key is empty", "setAttribute", cID[:], "S3_TAGS", pl, anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+	})
+
+	t.Run("value is empty", func(t *testing.T) {
+		var tags = map[string]string{
+			"k1": "value1",
+			"k2": "value2",
+			"k3": "",
+		}
+
+		pl, err := json.Marshal(tags)
+		require.NoError(t, err)
+
+		inv.InvokeFail(t, "tag k3 value is empty", "setAttribute", cID[:], "S3_TAGS", pl, anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+	})
+}
+
+func TestSetS3Attributes(t *testing.T) {
+	const anyValidUntil = math.MaxUint32
+	anyValidInvocScript := randomBytes(10)
+	anyValidVerifScript := randomBytes(10)
+	anyValidSessionToken := randomBytes(10)
+
+	blockChain, committee := chain.NewSingle(t)
+	require.Implements(t, (*neotest.MultiSigner)(nil), committee)
+	exec := neotest.NewExecutor(t, blockChain, committee, committee)
+
+	deployDefaultNNS(t, exec)
+	netmapContract := deployNetmapContract(t, exec, "ContainerFee", 0)
+	containerContract := neotest.CompileFile(t, exec.CommitteeHash, containerPath, path.Join(containerPath, "config.yml"))
+	deployBalanceContract(t, exec, netmapContract, containerContract.Hash)
+	deployProxyContract(t, exec)
+
+	exec.DeployContract(t, containerContract, nil)
+
+	owner := exec.NewAccount(t, 200_0000_0000)
+	ownerAcc := owner.ScriptHash()
+	ownerAddr := user.NewFromScriptHash(ownerAcc)
+	cnr := containertest.Container()
+	cnr.SetOwner(ownerAddr)
+	cnrBytes := cnr.Marshal()
+	cID := cid.NewFromMarshalledContainer(cnrBytes)
+	inv := exec.CommitteeInvoker(containerContract.Hash)
+
+	t.Run("create container", func(t *testing.T) {
+		const anyValidDomainName = ""
+		const anyValidDomainZone = ""
+
+		committeeInvoker := exec.CommitteeInvoker(containerContract.Hash)
+
+		_ = committeeInvoker.Invoke(t, stackitem.Null{}, "create",
+			cnrBytes, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken, anyValidDomainName, anyValidDomainZone, false)
+
+		assertGetInfo(t, inv, cID, cnr)
+	})
+
+	t.Run("store and remove S3_SETTINGS", func(t *testing.T) {
+		pl := "{}"
+
+		var cnr2 container.Container
+		cnr.CopyTo(&cnr2)
+
+		inv.Invoke(t, nil, "setAttribute", cID[:], "S3_SETTINGS", pl, anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+		cnr2.SetAttribute("S3_SETTINGS", pl)
+		assertGetInfo(t, inv, cID, cnr2)
+
+		inv.Invoke(t, nil, "removeAttribute", cID[:], "S3_SETTINGS", anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+		assertGetInfo(t, inv, cID, cnr)
+	})
+
+	t.Run("store and remove S3_NOTIFICATIONS", func(t *testing.T) {
+		pl := "{}"
+
+		var cnr2 container.Container
+		cnr.CopyTo(&cnr2)
+
+		inv.Invoke(t, nil, "setAttribute", cID[:], "S3_NOTIFICATIONS", pl, anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+		cnr2.SetAttribute("S3_NOTIFICATIONS", pl)
+		assertGetInfo(t, inv, cID, cnr2)
+
+		inv.Invoke(t, nil, "removeAttribute", cID[:], "S3_NOTIFICATIONS", anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+		assertGetInfo(t, inv, cID, cnr)
+	})
+
+	t.Run("invalid S3_SETTINGS", func(t *testing.T) {
+		pl := "invalid JSON"
+		inv.InvokeFail(t, "", "setAttribute", cID[:], "S3_SETTINGS", pl, anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
+	})
+
+	t.Run("invalid S3_NOTIFICATIONS", func(t *testing.T) {
+		pl := "invalid JSON"
+		inv.InvokeFail(t, "", "setAttribute", cID[:], "S3_NOTIFICATIONS", pl, anyValidUntil, anyValidInvocScript, anyValidVerifScript, anyValidSessionToken)
 	})
 }